refactor(integration): Integration test suite reorganized
