+++ zio-http/src/it/scala/zhttp/IntegrationSpec.scala
@@ -1,10 +1,7 @@
 package zhttp
 
 import zhttp.http._
-import zhttp.service.Client
 import zio._
-import zio.test.Assertion._
-import zio.test._
 
 object IntegrationSpec extends IntegrationRunnableSpec(80) {
   Runtime.default.unsafeRun(serve(app).forkDaemon)
@@ -15,22 +12,6 @@ object IntegrationSpec extends IntegrationRunnableSpec(80) {
   }
 
   def spec = suite("IntegrationSpec") {
-    HttpSpec
+    HttpIntegrationSpec.testSuite(addr, port)
   } provideCustomLayer env
-
-  def HttpSpec = suite("HttpSpec") {
-    testM("200 ok on /") {
-      val response = Client.request(s"http://${addr}:${port}")
-
-      assertM(response.map(_.status))(
-        equalTo(Status.OK),
-      )
-    } + testM("201 created on /post") {
-      val response = Client.request(
-        Client.ClientParams((Method.POST, URL(Path.apply(), URL.Location.Absolute(Scheme.HTTP, addr, port)))),
-      )
-
-      assertM(response.map(_.status))(equalTo(Status.CREATED))
-    }
-  }
 }
